+++ graphics/slider/slider.hpp
@@ -4,39 +4,39 @@
 
 class pGraphics::pSlider : public pGraphics::pElement {
     private:
-        std::pair<double, double> minMax;
+        std::pair<float, float> minMax;
         int decimals;
         int valueTextSize;
         pColor onColor;
         pColor offColor;
-        void(*function)(double value);
+        void(*function)(float value);
 
-        double value;
-        double pxOn;
-        double pxOff;
+        float value;
+        float pxOn;
+        float pxOff;
         std::string valueText;
 
         pGraphics::pText* valueTextObj;
 
     public:
-        pSlider(pGraphics* gfx, std::pair<double, double> pos, std::pair<double, double> size, std::pair<double, double> minMax, int decimals, std::string fontLocation, int valueTextSize, pColor onColor, pColor offColor, pColor valueTextColor, void(*function)(double));
+        pSlider(pGraphics* gfx, std::pair<float, float> pos, std::pair<float, float> size, std::pair<float, float> minMax, int decimals, std::string fontLocation, int valueTextSize, pColor onColor, pColor offColor, pColor valueTextColor, void(*function)(float));
         ~pSlider();
 
         void init();
         void draw();
         void handleMouse();
 
-        void setPos(std::pair<double, double> newPos);
-        void setSize(std::pair<double, double> newSize);
-        void setMinMax(std::pair<double, double> newMinMax);
+        void setPos(std::pair<float, float> newPos);
+        void setSize(std::pair<float, float> newSize);
+        void setMinMax(std::pair<float, float> newMinMax);
         void setDecimals(int newDecimals);
         void setOnColor(pColor newOnColor) { onColor = newOnColor; };
         void setOffColor(pColor newOffColor) { offColor = newOffColor; };
-        void setFunction(void(*newFunction)(double)) { function = newFunction; };
-        void setValue(double newValue);
+        void setFunction(void(*newFunction)(float)) { function = newFunction; };
+        void setValue(float newValue);
         
-        std::pair<double, double> getPos() { return this->pos; };
-        std::pair<double, double> getSize() { return this->size; };
+        std::pair<float, float> getPos() { return this->pos; };
+        std::pair<float, float> getSize() { return this->size; };
         pGraphics::pText* getValueTextObj() { return valueTextObj; };
-        double getValue() { return value; };
+        float getValue() { return value; };
 };
